tp bank end

+++ Banque/src/main/java/org/example/dao/BaseDao.java
@@ -1,5 +1,7 @@
 package org.example.dao;
 
+import org.example.entity.Account;
+
 import java.sql.Connection;
 import java.sql.PreparedStatement;
 import java.sql.ResultSet;
@@ -20,7 +22,8 @@ public abstract class BaseDao<T> {
     public abstract boolean update(T element) throws SQLException;
     public abstract boolean delete(T element) throws SQLException;
 
-    public abstract T get(int id) throws SQLException;
+    public abstract T get(long id) throws SQLException;
+
 
     public abstract List<T> get() throws SQLException;
 }
